Retire arviss_impl.h.
