+++ arviss.c
@@ -1,6 +1,5 @@
 #include "arviss.h"
 
-#include "arviss_impl.h"
 #include "conversions.h"
 #include "opcodes.h"
