+++ tests/CMakeLists.txt
@@ -1,4 +1,4 @@
 # Arviss tests.
-add_executable(decode_test decode_test.cpp ../arviss_impl.h)
+add_executable(decode_test decode_test.cpp ../arviss.h)
 target_link_libraries(decode_test PRIVATE arviss gtest_main)
 add_test(decode_test decode_test)
